re2: more -Wno-sign-compare fixes
Reviewed-on: https://code-review.googlesource.com/1274

+++ re2/testing/regexp_test.cc
@@ -29,9 +29,9 @@ TEST(Regexp, BigConcat) {
   Regexp* x;
   x = Regexp::Parse("x", Regexp::NoParseFlags, NULL);
   vector<Regexp*> v(90000, x);  // ToString bails out at 100000
-  for (int i = 0; i < v.size(); i++)
+  for (size_t i = 0; i < v.size(); i++)
     x->Incref();
-  CHECK_EQ(x->Ref(), 1 + v.size()) << x->Ref();
+  CHECK_EQ(x->Ref(), 1 + static_cast<int>(v.size())) << x->Ref();
   Regexp* re = Regexp::Concat(&v[0], v.size(), Regexp::NoParseFlags);
   CHECK_EQ(re->ToString(), string(v.size(), 'x'));
   re->Decref();
