+++ re2/testing/string_generator.cc
@@ -44,13 +44,13 @@ void StringGenerator::Reset() {
 bool StringGenerator::IncrementDigits() {
   // First try to increment the current number.
   for (int i = digits_.size() - 1; i >= 0; i--) {
-    if (++digits_[i] < alphabet_.size())
+    if (++digits_[i] < static_cast<int>(alphabet_.size()))
       return true;
     digits_[i] = 0;
   }
 
   // If that failed, make a longer number.
-  if (digits_.size() < maxlen_) {
+  if (digits_.size() < static_cast<size_t>(maxlen_)) {
     digits_.push_back(0);
     return true;
   }
@@ -84,7 +84,7 @@ const StringPiece& StringGenerator::Next() {
     return sp_;
   }
   s_.clear();
-  for (int i = 0; i < digits_.size(); i++) {
+  for (size_t i = 0; i < digits_.size(); i++) {
     s_ += alphabet_[digits_[i]];
   }
   hasnext_ = random_ ? RandomDigits() : IncrementDigits();
